requirements.txt
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
colorama==0.4.6
colorlog==6.10.1
constantly==23.10.4
cryptography==46.0.4
daphne==4.2.1
dill==0.4.1
Django==6.0.1
exceptiongroup==1.3.1
fastjsonschema==2.21.2
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jsonpatch==1.33
jsonpointer==3.0.0
lxml==6.0.2
msgpack==1.1.2
mypy_extensions==1.1.0
nest-asyncio==1.6.0
orjson==3.11.6
packaging==26.0
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyOpenSSL==25.3.0
python-dotenv==1.2.1
reactpy==1.1.0
reactpy_django==5.2.1
reactpy_router==1.0.3
requests==2.32.5
service-identity==24.2.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
zope.interface==8.2
